Revert to buffered response
The http handler closes the requests body after the first response flush. Ideally it won't close the request body and after each request body line read we will write a response in a streaming fashion, thus not consuming tons of RAM. This will result in "Transfer-Encoding: chunked" instead of buffering the response and sending "Content-Length: XXX".